+++ package.json
@@ -9,6 +9,7 @@
     "electron": "^2.0.2"
   },
   "dependencies": {
-    "electron-json-storage": "^4.1.0"
+    "electron-json-storage": "^4.1.0",
+    "electron-json-storage-sync": "^1.1.1"
   }
 }
